Corrected PersonDto.personId from int to long

+++ src/main/java/com/alienworkspace/cdr/model/dto/person/PersonDto.java
@@ -9,7 +9,7 @@ import java.time.LocalDate;
 @Builder
 public class PersonDto {
 
-    private int personId;
+    private long personId;
 
     private Character gender;
